pyproject.toml
[project]
name = "multi-agent-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.135.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.3.0",
    "torch==2.9.1",
    "torchvision==0.24.1",
    "torchaudio==2.9.1",
    "transformers>=5.3.0",
    "uvicorn>=0.42.0",
    "asyncpg>=0.31.0",
    "langchain-text-splitters>=1.1.1",
    "pypdf>=6.9.2",
    "python-docx>=1.2.0",
]

[tool.uv.sources]
torch = { index = "pytorch-cuda" }
torchvision = { index = "pytorch-cuda" }
torchaudio = { index = "pytorch-cuda" }

[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu130"  
explicit = true
